feat(ast_builder): move to enum FunctionNode
